refactor fuse_with_classif and merge_label_list to utils
