+++ tests/test_utils.py
@@ -1,7 +1,7 @@
 import pytest
 import numpy as np
 import numpy.testing as npt
-from tinnsleep.utils import epoch, compute_nb_epochs
+from tinnsleep.utils import epoch, compute_nb_epochs, merge_labels_list, fuse_with_classif_result
 
 
 def test_compute_nb_epochs():
@@ -82,3 +82,45 @@ def test_epoch_fail_size():
         Ne = 8  # electrodes
         X = np.random.randn(Ne, N)
         epochs = epoch(X.T, T, I, axis=0)
+
+
+def test_merge_labels_list():
+
+    # Unchanging a list to the same number of elements
+    v_lab = merge_labels_list([[True, False, True, False, True]], 5)
+    npt.assert_equal(v_lab, [True, False, True, False, True])
+
+    # Unchanging two identical list into one:
+    v_lab = merge_labels_list([[True, False, True, False, True], [True, False, True, False, True]], 5)
+    npt.assert_equal(v_lab, [True, False, True, False, True])
+
+    # dealing with 2 coherent arrays of len(l) and 2*len(l) and getting output in 2*len(l):
+    v_lab = merge_labels_list([[True, False], [True, True, False, False]], 4)
+    npt.assert_equal(v_lab, [True, True, False, False])
+
+    # dealing with classic situation:
+    v_lab = merge_labels_list([[True, False], [True, True, True, True]], 4)
+    npt.assert_equal(v_lab, [True, True, False, False])
+
+    # dealing with classic situation 2:
+    v_lab = merge_labels_list([[True, True], [True, True, False, True]], 4)
+    npt.assert_equal(v_lab, [True, True, False, True])
+
+    # dealing with 2 coherent arrays of len(l) and 2*len(l) and getting output in len(l):
+    v_lab = merge_labels_list([[True, False], [True, True, False, False]], 2)
+    npt.assert_equal(v_lab, [True, False])
+
+    # dealing with tricky case 1:
+    v_lab = merge_labels_list([[True, True], [True, True, True, False]], 2)
+    npt.assert_equal(v_lab, [True, False])
+
+    # dealing with tricky case 2:
+    v_lab = merge_labels_list([[True, True, False, True], [True, True]], 2)
+    npt.assert_equal(v_lab, [True, False])
+
+
+def test_fuse_with_classif_result():
+    check_imp = [[False, False], [False, True], [True, True], [True, False], [True, True], [True, False]]
+    classif=np.asanyarray([1, 2, 3, 4])
+    classif = fuse_with_classif_result(check_imp, classif)
+    npt.assert_equal(classif, [1, 2, False, 3, False, 4])
